requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
pypdf>=3.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
requests>=2.31.0
